Implement calculation of average cost in USD
Implemented calculation of average cost in USD for the output portfolio.
This is useful when the average cost method is not configured for
displaying cost basis on the portfolio page. IBKR defaults to "statement
cost" which means the average price shown will sync up with your
statement cost basis overnight, using the selected lot matching method
(which defaults to FIFO).

Note that the average cost method is required by Swedish tax law for
calculating profit and loss.